+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Increased visibility of not-master-branch warning
 * Base java dockerfile updated to Gradle 3.2.1
 * Changed `acceptance-tests` alias to take in a parameter specifying the container name. All other parameters are passed through to the run_tests.sh script. This is so the skeleton version of that script can in turn pass them on to cucumber. Also added an identical, shorter `acctest` alias.
+* Created a v2 of base python and base flask dockerfiles. Changes are the removal of the deprecated FLASK_LOG_LEVEL env var and the removal of the LC_ALL env var. Be aware when migrating your apps that this will break installation of httpretty.
 
 ## v1.2.0