+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.2.1
+
+* Pin version of vagrant-persistent-storage plugin to fix IDE controller bug
+* Fixed db2 alias
+* Increased visibility of not-master-branch warning
+* Base java dockerfile updated to Gradle 3.2.1
+* Changed `acceptance-tests` alias to take in a parameter specifying the container name. All other parameters are passed through to the run_tests.sh script. This is so the skeleton version of that script can in turn pass them on to cucumber. Also added an identical, shorter `acctest` alias.
+* Created a v2 of base python and base flask dockerfiles. Changes are the removal of the deprecated FLASK_LOG_LEVEL env var and the removal of the LC_ALL env var. Be aware when migrating your apps that this will break installation of httpretty.
+
 ## v1.2.0
 
 * Updated Ruby/gems/bundler versions in base ruby dockerfile